refactor: extract interaction logic of GamePageList to new ViewModel
